(SIMP-1169) Add RuboCop, fix for failures
This adds RuboCop and its Rake task to this repo, as well as to the
baseline modulesync configs, and fixes all RuboCop failures.
Additionally, for this repo, it adds RuboCop as a Travis-CI test.

A few minor changes were made to the RuboCop configuration.  Namely,
changes were made to maintain compatability with Ruby 1.8.7, because we
need to maintain compatability with RHEL/CentOS 5 & 6 until they reach
EOL.  A few other changes are set to adhere to the Puppet Style Guide
where the are conflicts-- for example, line length and end-of-line
commas.