Workaround for Hocon::Impl::TokenType.name issue
This works around a break in virtually all ServerSpec tests, where each
test returns an `ArgumentError` with the message `wrong number of
arguments (0 for 1)`.  The workaround is to lock Specinfra to a pre-2.59
version, although the bug is in HOCON.

See https://github.com/puppetlabs/ruby-hocon/issues/75 for more details.